Finished Membership Plan Backend CRUD

+++ gym-backend/src/main/java/com/gymsystem/gymbackend/repository/MembershipPlanRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MembershipPlanRepository extends JpaRepository<MembershipPlan, Integer> {
+public interface MembershipPlanRepository extends JpaRepository<MembershipPlan, Long> {
 }
